requirements.txt
mysql_connector_repackaged==0.3.1
prompt_toolkit==3.0.30
selenium==4.4.3
stdiomask==0.0.6
webdriver_manager==3.8.3
